+++ app/assets/stylesheets/custom.scss.erb
@@ -4886,3 +4886,26 @@ float: right;
 }
 
 
+#user-login-history-modal {
+  ul {
+    list-style-type: none;
+  }
+
+  li {
+    padding-top: 3%;
+    span.success {
+      color: green;
+    }
+    span.failure {
+      color: red;
+    }
+  }
+
+  ul.pagination {
+    background: $mhc-dark-blue;
+
+    li a {
+      background: $mhc-dark-blue;
+    }
+  }
+}
