+++ app/controllers/application_controller.rb
@@ -224,7 +224,7 @@ class ApplicationController < ActionController::Base
       SAVEUSER[:current_user_id] = current_user.try(:id)
       session_id = SessionTaggedLogger.extract_session_id_from_request(request)
       unless SessionIdHistory.where(session_id: session_id).present?
-        SessionIdHistory.create(session_id: session_id, session_user_id: current_user.try(:id))
+        SessionIdHistory.create(session_id: session_id, session_user_id: current_user.try(:id), sign_in_outcome: "Successful", ip_address: request.remote_ip)
       end
     end
