+++ app/controllers/users/passwords_controller.rb
@@ -1,7 +1,19 @@
 class Users::PasswordsController < Devise::PasswordsController
   before_filter :confirm_identity, only: [:create]
   def create
-    super
+    self.resource = resource_class.send_reset_password_instructions(resource_params)
+    yield resource if block_given?
+    if successfully_sent?(resource)
+      resource.security_question_responses.destroy_all
+      
+      respond_to do |format|
+       format.html { respond_with({}, location: after_sending_reset_password_instructions_path_for(resource_name)) }
+       format.js
+      end
+
+    else
+      respond_with(resource)
+    end
   end
 
   private
@@ -11,6 +23,7 @@ class Users::PasswordsController < Devise::PasswordsController
   end
 
   def confirm_identity
+    return true if current_user.has_role? 'hbx_staff'
     unless user.identity_confirmed_token == params[:user][:identity_confirmed_token]
       flash[:error] = "Something went wrong, please try again"
       redirect_to new_user_password_path
