+++ app/models/session_id_history.rb
@@ -4,8 +4,15 @@ class SessionIdHistory
 
   field :session_user_id,  type: String
   field :session_id, type: String
+  field :sign_in_outcome, type: String
+  field :ip_address, type: String
 
   index({ session_user_id: 1})
   index({ session_id: 1})
-end 
-  
+
+  class << self
+    def for_user(user_id:)
+      self.where(session_user_id: user_id)
+    end
+  end
+end
