+++ app/models/user.rb
@@ -2,32 +2,19 @@ class User
   INTERACTIVE_IDENTITY_VERIFICATION_SUCCESS_CODE = "acc"
   MIN_USERNAME_LENGTH = 8
   MAX_USERNAME_LENGTH = 60
-  MAX_SAME_CHAR_LIMIT = 4
+
   include Mongoid::Document
   include Mongoid::Timestamps
   include Acapi::Notifiers
-
+  include AuthorizationConcern
+  include PermissionsConcern
   attr_accessor :login
-
-  # Include default devise modules. Others available are:
-  # :confirmable, :lockable, :timeoutable and :omniauthable
-  devise :database_authenticatable, :registerable, :lockable,
-         :recoverable, :rememberable, :trackable, :timeoutable, :authentication_keys => {email: false, login: true}
-
-  embeds_many :security_question_responses
 
   validates_presence_of :oim_id
   validates_uniqueness_of :oim_id, :case_sensitive => false
-  validate :password_complexity
   validate :oim_id_rules
   validates_uniqueness_of :email,:case_sensitive => false
-  validates_presence_of     :password, if: :password_required?
-  validates_confirmation_of :password, if: :password_required?
-  validates_length_of       :password, within: Devise.password_length, allow_blank: true
-  validates_format_of :email, with: Devise::email_regexp , allow_blank: true, :message => "(optional) is invalid"
 
-  scope :locked, ->{ where(:locked_at.ne => nil) }
-  scope :unlocked, ->{ where(locked_at: nil) }
   scope :datatable_search, ->(query) { self.where({"$or" => ([{"name" => Regexp.compile(Regexp.escape(query), true)}, {"email" => Regexp.compile(Regexp.escape(query), true)}])}) }
 
   def oim_id_rules
@@ -40,57 +27,9 @@ class User
     end
   end
 
-  def password_complexity
-    if password.present? and not password.match(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^a-zA-Z\d ]).+$/)
-      errors.add :password, "must include at least one lowercase letter, one uppercase letter, one digit, and one character that is not a digit or letter or space"
-    elsif password.present? and password.match(/#{Regexp.escape(oim_id)}/i)
-      errors.add :password, "cannot contain username"
-    elsif password.present? and password_repeated_chars_limit(password)
-      errors.add :password, "cannot repeat any character more than #{MAX_SAME_CHAR_LIMIT} times"
-    elsif password.present? and password.match(/(.)\1\1/)
-      errors.add :password, "must not repeat consecutive characters more than once"
-    elsif password.present? and !password.match(/(.*?[a-zA-Z]){4,}/)
-      errors.add :password, "must have at least 4 alphabetical characters"
-    end
-  end
-
-  def password_repeated_chars_limit(password)
-    return true if password.chars.group_by(&:chr).map{ |k,v| v.size}.max > MAX_SAME_CHAR_LIMIT
-    false
-  end
-
-  def password_required?
-    !persisted? || !password.nil? || !password_confirmation.nil?
-  end
-
   def valid_attribute?(attribute_name)
     self.valid?
     self.errors[attribute_name].blank?
-  end
-
-  def self.password_invalid?(password)
-    user = User.new(oim_id: 'example1', password: password)
-    !user.valid_attribute?('password')
-  end
-
-  def self.generate_valid_password
-    password = Devise.friendly_token.first(16)
-    password = password + "aA1!"
-    password = password.squeeze
-    if password_invalid?(password)
-      password = generate_valid_password
-    else
-      password
-    end
-  end
-
-  def self.find_for_database_authentication(warden_conditions)
-    conditions = warden_conditions.dup
-    if login = conditions.delete(:login).downcase
-      where(conditions).where('$or' => [ {:oim_id => /^#{Regexp.escape(login)}$/i}, {:email => /^#{Regexp.escape(login)}$/i} ]).first
-    else
-      where(conditions).first
-    end
   end
 
   def switch_to_idp!
@@ -118,15 +57,6 @@ class User
   ## Seed: https://github.com/plataformatec/devise/wiki/How-To%3a-Require-admin-to-activate-account-before-sign_in
   field :approved, type: Boolean, default: true
 
-  ## Database authenticatable
-  field :email,              type: String, default: ""
-  field :encrypted_password, type: String, default: ""
-
-  ## Recoverable
-  field :reset_password_token,   type: String
-  field :reset_password_sent_at, type: Time
-  field :identity_confirmed_token, type: String
-
   ##RIDP
   field :identity_verified_date, type: Date
   field :identity_final_decision_code, type: String
@@ -134,18 +64,9 @@ class User
   field :identity_response_code, type: String
   field :identity_response_description_text, type: String
 
-  ## Rememberable
-  field :remember_created_at, type: Time
-
   ## Trackable
-  field :sign_in_count,      type: Integer, default: 0
-  field :current_sign_in_at, type: Time
-  field :last_sign_in_at,    type: Time
-  field :current_sign_in_ip, type: String
-  field :last_sign_in_ip,    type: String
   field :idp_uuid, type: String
 
-  field :authentication_token
   field :roles, :type => Array, :default => []
 
   # Oracle Identity Manager ID
@@ -154,11 +75,6 @@ class User
   field :last_portal_visited, type: String
   field :idp_verified, type: Boolean, default: false
 
-  ## Lockable
-  field :failed_attempts, type: Integer, default: 0 # Only if lock strategy is :failed_attempts
-  field :unlock_token,    type: String # Only if unlock strategy is :email or :both
-  field :locked_at,       type: Time
-
   index({preferred_language: 1})
   index({approved: 1})
   index({roles: 1},  {sparse: true}) # MongoDB multikey index
@@ -166,48 +82,10 @@ class User
   index({oim_id: 1}, {sparse: true, unique: true})
   index({created_at: 1 })
 
-
   before_save :strip_empty_fields
-
-  ROLES = {
-    employee: "employee",
-    resident: "resident",
-    consumer: "consumer",
-    broker: "broker",
-    system_service: "system_service",
-    web_service: "web_service",
-    hbx_staff: "hbx_staff",
-    employer_staff: "employer_staff",
-    broker_agency_staff: "broker_agency_staff",
-    general_agency_staff: "general_agency_staff",
-    assister: 'assister',
-    csr: 'csr',
-  }
 
   # Enable polymorphic associations
   belongs_to :profile, polymorphic: true
-
-  has_one :person
-  accepts_nested_attributes_for :person, :allow_destroy => true
-
-  # after_initialize :instantiate_person
-
-  ## Confirmable
-  # field :confirmation_token,   type: String
-  # field :confirmed_at,         type: Time
-  # field :confirmation_sent_at, type: Time
-  # field :unconfirmed_email,    type: String # Only if using reconfirmable
-
-  ## Lockable
-  field :failed_attempts, type: Integer, default: 0 # Only if lock strategy is :failed_attempts
-  field :unlock_token,    type: String # Only if unlock strategy is :email or :both
-  field :locked_at,       type: Time
-
-  before_save :ensure_authentication_token
-
-  #  after_create :send_welcome_email
-
-  delegate :primary_family, to: :person, allow_nil: true
 
   attr_accessor :invitation_id
   #  validate :ensure_valid_invitation, :on => :create
@@ -228,11 +106,6 @@ class User
     end
   end
 
-  def person_id
-    return nil unless person.present?
-    person.id
-  end
-
   def idp_verified?
     idp_verified
   end
@@ -240,93 +113,6 @@ class User
   def send_welcome_email
     UserMailer.welcome(self).deliver_now
     true
-  end
-
-  def set_random_password(passwd)
-    self.password = passwd
-    self.password_confirmation = passwd
-    self.save
-  end
-
-  def update_lockable
-    if locked_at.nil?
-      self.lock_access!
-    else
-      self.unlock_access!
-    end
-  end
-
-  def lockable_notice
-    self.locked_at.nil? ? 'unlocked' : 'locked'
-  end
-
-  def has_role?(role_sym)
-    return false if person_id.blank?
-    roles.any? { |r| r == role_sym.to_s }
-  end
-
-  def has_employee_role?
-    person && person.active_employee_roles.present?
-  end
-
-  def has_consumer_role?
-    person && person.consumer_role
-  end
-
-  def has_resident_role?
-    person && person.resident_role
-  end
-
-  def has_employer_staff_role?
-    person && person.has_active_employer_staff_role?
-  end
-
-  def has_broker_agency_staff_role?
-    has_role?(:broker_agency_staff)
-  end
-
-  def has_general_agency_staff_role?
-    has_role?(:general_agency_staff)
-  end
-
-  def has_insured_role?
-    has_employee_role? || has_consumer_role?
-  end
-
-  def has_broker_role?
-    has_role?(:broker)
-  end
-
-  def has_hbx_staff_role?
-    has_role?(:hbx_staff) || self.try(:person).try(:hbx_staff_role)
-  end
-
-  def has_csr_role?
-    has_role?(:csr)
-  end
-
-  def has_csr_subrole?
-    person && person.csr_role && !person.csr_role.cac
-  end
-
-  def has_cac_subrole?
-    person && person.csr_role && person.csr_role.cac
-  end
-
-  def has_assister_role?
-    has_role?(:assister)
-  end
-
-  def has_agent_role?
-    has_role?(:csr) || has_role?(:assister)
-  end
-
-  def can_change_broker?
-    if has_employer_staff_role? || has_hbx_staff_role?
-      true
-    elsif has_general_agency_staff_role? || has_broker_role? || has_broker_agency_staff_role?
-      false
-    end
   end
 
   def agent_title
@@ -345,84 +131,12 @@ class User
     person == employer_profile.active_broker if employer_profile.active_broker
   end
 
-  def ensure_authentication_token
-    if authentication_token.blank?
-      self.authentication_token = generate_authentication_token
-    end
-    true
-  end
-
-  def instantiate_person
-    self.person = Person.new
-  end
-
-  # Instances without a matching Person model
-  # This suboptimal query approach is necessary, as the belongs_to side of the association holds the
-  #   ID in a has_one association
-  def self.orphans
-    user_ids = Person.where(:user_id => { "$ne" => nil }).pluck(:user_id)
-    User.where("_id" => { "$nin" => user_ids }).order(email: :asc).entries
-  end
-
-  def self.send_reset_password_instructions(attributes={})
-    recoverable = find_or_initialize_with_errors(reset_password_keys, attributes, :not_found)
-    if !recoverable.approved?
-      recoverable.errors[:base] << I18n.t("devise.failure.not_approved")
-    elsif recoverable.persisted?
-      recoverable.send_reset_password_instructions
-    end
-    recoverable
-  end
-
-  def self.find_by_authentication_token(token)
-    where(authentication_token: token).first
-  end
-
   def handle_headless_records
     headless_with_email = User.where(email: /^#{Regexp.quote(email)}$/i)
     headless_with_oim_id = User.where(oim_id: /^#{Regexp.quote(oim_id)}$/i)
     headless_users = headless_with_email + headless_with_oim_id
     headless_users.each do |headless|
       headless.destroy if !headless.person.present?
-    end
-  end
-
-  def needs_to_provide_security_questions?
-    security_question_responses.length < 3
-  end
-
-  class << self
-
-    def by_email(email)
-      where(email: /^#{email}$/i).first
-    end
-
-    def current_user=(user)
-      Thread.current[:current_user] = user
-    end
-
-    def logins_before_captcha
-      4
-    end
-
-    def login_captcha_required?(login)
-      begin
-        logins_before_captcha <= self.or({oim_id: login}, {email: login}).first.failed_attempts
-      rescue => e
-        true
-      end
-    end
-
-    def current_user
-      Thread.current[:current_user]
-    end
-
-    def logins_before_captcha
-      4
-    end
-
-    def has_answered_question? security_question_id
-      where(:'security_question_responses.security_question_id' => security_question_id).any?
     end
   end
 
@@ -479,36 +193,70 @@ class User
     announcements.uniq
   end
 
-  def self.get_saml_settings
-    settings = OneLogin::RubySaml::Settings.new
+  class << self
+    def password_invalid?(password)
+      ## TODO: oim_id is an explicit dependency to the User class
+      resource = self.new(oim_id: 'example1', password: password)
+      !resource.valid_attribute?('password')
+    end
 
-    # When disabled, saml validation errors will raise an exception.
-    settings.soft = true
+    def find_for_database_authentication(warden_conditions)
+      #TODO: Another explicit oim_id dependency
+      conditions = warden_conditions.dup
+      if login = conditions.delete(:login).downcase
+        where(conditions).where('$or' => [ {:oim_id => /^#{Regexp.escape(login)}$/i}, {:email => /^#{Regexp.escape(login)}$/i} ]).first
+      else
+        where(conditions).first
+      end
+    end
 
-    # SP section
-    settings.assertion_consumer_service_url = SamlInformation.assertion_consumer_service_url
-    settings.assertion_consumer_logout_service_url = SamlInformation.assertion_consumer_logout_service_url
-    settings.issuer                         = SamlInformation.issuer
+    def by_email(email)
+      where(email: /^#{email}$/i).first
+    end
 
-    # IdP section
-    settings.idp_entity_id                  = SamlInformation.idp_entity_id
-    settings.idp_sso_target_url             = SamlInformation.idp_sso_target_url
-    settings.idp_slo_target_url             = SamlInformation.idp_slo_target_url
-    settings.idp_cert                       = SamlInformation.idp_cert
-    # or settings.idp_cert_fingerprint           = "3B:05:BE:0A:EC:84:CC:D4:75:97:B3:A2:22:AC:56:21:44:EF:59:E6"
-    #    settings.idp_cert_fingerprint_algorithm = XMLSecurity::Document::SHA1
+    def current_user=(user)
+      Thread.current[:current_user] = user
+    end
 
-    settings.name_identifier_format         = SamlInformation.name_identifier_format
+    def get_saml_settings
+      settings = OneLogin::RubySaml::Settings.new
 
-    # Security section
-    settings.security[:authn_requests_signed] = false
-    settings.security[:logout_requests_signed] = false
-    settings.security[:logout_responses_signed] = false
-    settings.security[:metadata_signed] = false
-    settings.security[:digest_method] = XMLSecurity::Document::SHA1
-    settings.security[:signature_method] = XMLSecurity::Document::RSA_SHA1
+      # When disabled, saml validation errors will raise an exception.
+      settings.soft = true
 
-    settings
+      # SP section
+      settings.assertion_consumer_service_url = SamlInformation.assertion_consumer_service_url
+      settings.assertion_consumer_logout_service_url = SamlInformation.assertion_consumer_logout_service_url
+      settings.issuer                         = SamlInformation.issuer
+
+      # IdP section
+      settings.idp_entity_id                  = SamlInformation.idp_entity_id
+      settings.idp_sso_target_url             = SamlInformation.idp_sso_target_url
+      settings.idp_slo_target_url             = SamlInformation.idp_slo_target_url
+      settings.idp_cert                       = SamlInformation.idp_cert
+      # or settings.idp_cert_fingerprint           = "3B:05:BE:0A:EC:84:CC:D4:75:97:B3:A2:22:AC:56:21:44:EF:59:E6"
+      #    settings.idp_cert_fingerprint_algorithm = XMLSecurity::Document::SHA1
+
+      settings.name_identifier_format         = SamlInformation.name_identifier_format
+
+      # Security section
+      settings.security[:authn_requests_signed] = false
+      settings.security[:logout_requests_signed] = false
+      settings.security[:logout_responses_signed] = false
+      settings.security[:metadata_signed] = false
+      settings.security[:digest_method] = XMLSecurity::Document::SHA1
+      settings.security[:signature_method] = XMLSecurity::Document::RSA_SHA1
+
+      settings
+    end
+
+    # Instances without a matching Person model
+    # This suboptimal query approach is necessary, as the belongs_to side of the association holds the
+    #   ID in a has_one association
+    def orphans
+      user_ids = Person.where(:user_id => { "$ne" => nil }).pluck(:user_id)
+      User.where("_id" => { "$nin" => user_ids }).order(email: :asc).entries
+    end
   end
 
   private
@@ -516,12 +264,5 @@ class User
   def strip_empty_fields
     unset("email") if email.blank?
     unset("oim_id") if oim_id.blank?
-  end
-
-  def generate_authentication_token
-    loop do
-      token = Devise.friendly_token
-      break token unless User.where(authentication_token: token).first
-    end
   end
 end
