+++ app/views/datatables/shared/_dropdown.html.erb
@@ -1,6 +1,6 @@
 <div class="dropdown pull-right" id="<%= row_actions_id %>">
-  <button class="btn btn-default dropdown-toggle" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" id="dropdown_for_<%= row_actions_id %>">
-    Actions 
+	<button class="btn btn-default dropdown-toggle" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" id="dropdown_for_<%= row_actions_id %>">
+    Actions
     <span class="caret"></span>
   </button>
   <ul class="dropdown-menu" aria-labelledby="dropdown_for_<%= row_actions_id %>">
