+++ app/views/devise/passwords/create.js.erb
@@ -1,0 +1,1 @@
+$('.content-bottom').prepend("<%= escape_javascript(render 'layouts/flash', message: 'User was emailed a password reset link', type: 'success') %>");
