+++ app/views/exchanges/hbx_profiles/shared/_primary_nav.html.erb
@@ -106,14 +106,12 @@
         <% end %>
       </li>
 
-      <% if individual_market_is_enabled? %>
       <li role="presentation"<%= menu_tab_class(active_tab, "user-account-tab") %> class="tab-second <%=pundit_class HbxProfile, :view_admin_tabs?%>">
         <%= link_to(user_account_index_exchanges_hbx_profiles_path, :remote => true, "aria-expanded" => "true", "aria-controls" => "users", "role" => "tab" ) do  %>
           <span class="hidden-xs">User Accounts</span>
           <span class="glyphicon glyphicon-th-list hidden-md hidden-sm hidden-lg" title="User Accounts"></span>
         <% end %>
       </li>
-      <% end %>
       <!-- <li role="presentation"<%= menu_tab_class(active_tab, "binder-tab") %> class="tab-second">
        <%= link_to(binder_index_exchanges_hbx_profiles_path, :remote => true, "aria-expanded" => "true", "aria-controls" => "binder", "role" => "tab" ) do  %>
         <span class="hidden-xs">Binder Transition</span>
