+++ app/views/shared/modals/_edit_email_form.html.erb
@@ -1,0 +1,28 @@
+<div class="modal fade" id="setUserEmailModal" tabindex="-1" role="dialog" aria-labelledby="setUserEmailModal">
+  <div class="modal-dialog" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h4 class="modal-title" id="setUserEmailModalLabel">
+          Please enter email for sending reset password instruction
+        </h4>
+      </div>
+      <%= form_for user, remote: true do |f| %>
+          <div class="modal-body">
+            <div class="modal-questions-container">
+              <div class="question-wrapper">
+                <div class="input-wrapper">
+                  <%= f.label :email %>
+                  <%= f.text_field :email, required: true %>
+                </div>
+              </div>
+            </div>
+          </div>
+          <div class="modal-footer">
+            <%= f.button 'Submit', class: "btn btn-danger" %>
+            <%= link_to 'Close', '', :class => 'btn btn-default', data: { dismiss: "modal"} %>
+          </div>
+          </div>
+      <% end %>
+
+      </div>
+</div>
