+++ app/views/users/_login_history.html.erb
@@ -1,0 +1,32 @@
+<div class="modal fade" id="user-login-history-modal" tabindex="-1" role="dialog">
+  <div class="modal-dialog" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+        <h4 class="modal-title"><%= "Login history for #{@user.email}" %></h4>
+      </div>
+      <div class="modal-body">
+        <ul>
+          <% login_history.each do |login| %>
+            <li>
+              <% if login.sign_in_outcome == 'Successful' %>
+                <span class="glyphicon glyphicon-ok-circle success"></span> Successful
+              <% else %>
+                <span class="glyphicon glyphicon-remove-circle failure"></span> Failed
+              <% end %>
+              login on: <%= TimeKeeper.local_time(login.created_at).strftime('%D %r') %> from: <%= login.ip_address || "IP NOT FOUND" %></li>
+          <% end %>
+        </ul>
+        <div class="row text-center">
+          <%= paginate login_history, remote: true %>
+        </div>
+        <div class="row text-center">
+          <%= page_entries_info login_history, entry_name: 'login records' %>
+        </div>
+      </div>
+      <div class="modal-footer">
+        <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
+      </div>
+    </div><!-- /.modal-content -->
+  </div><!-- /.modal-dialog -->
+</div><!-- /.modal -->
