+++ app/views/users/login_history.js.erb
@@ -1,0 +1,6 @@
+if ($("#user-login-history-modal").length == 0) {
+  $('.tab-content').append("<%=escape_javascript(render('login_history', login_history: @user_login_history, user: @user)) %>");
+} else {
+  $('#user-login-history-modal').replaceWith("<%=escape_javascript(render('login_history', login_history: @user_login_history)) %>");
+}
+$('#user-login-history-modal').modal('show');
